Finally got the pathing working
Generated a mock path using geojson.io.  This emulates a car path which drives around my neighborhood.  Using leaflet moving marker I was able to set this in motion.  The hold up was the default structure of the cars.  it needed to be an array of objects BEFORE importing.  Not sure why. dont care why.