reworked merc form
client info now properly populates in the merc form.  this needed to be changed from formSchema to basic bitch v-models to save some time and remove some unneccesary complexity.
this also makes adding these details to the popup easier.
Oh and the popup now properly displays properties which are sent to it.