google auth is now wired up on mobile.
There is still a small discrepancy between  the app-dev and app-prod SHA keys.  This is causing ONLY prod to work locally, which is fine for now but i would really like dev to be working under it's own hash so i dont have to overwrite the app when deving local.